rename: 채팅방 이벤트 리스너, 채팅 유저 이벤트 리스너 애플리케이션 패키지로 이동 #248
